requirements.txt
hsms
numpy
picamera
picamera2
Pillow==8.2.0
six==1.16.0
qrcode
spidev==3.5
pyzbar
RPi.GPIO==0.7.1
pycrypto
